Restructure for multiple packages.
