+++ docs/.cspell/workspace.txt
@@ -6,47 +6,107 @@ aiohttp
 aiopath
 amodio
 anyio
+autorig
+backgroundrenderer
+bakeanimation
+bvhviewer
+canvaseventtypes
 celerybeat
+channelwranglesnippet
+charactertoolutils
+choptoolutils
+cloneconfigdialog
+cloudsubmit
+cloudtoolutils
+coloraide
+colorschemeutils
 Connor
+coptoolutils
+crowdstoolutils
+curveutils
 cython
 davidhewitt
 dbaeumer
 declutters
+defaultstatetools
+defaulttoolmenu
+defaulttools
 devel
+digitalassetsupport
 direnv
 dmypy
 dockerized
+dopclothproxy
+dopclothtoolutils
+dopfemtoolutils
+dopgeofiltertoolutils
+dopinstance
+dopparticlefluidtoolutils
+doppoptoolutils
+doppyrotoolutils
+doprbdtoolutils
+dopsmoketoolutils
+dopsparsepyrotools
+dopstatictoolutils
+doptoolutils
+dopwiretoolutils
 dotenv
+dragdroputils
+drivertoolutils
 envrc
 esbuild
 eslintcache
 excs
+expressionmenu
 fanout
 fastapi
 fastio
 fastmcp
+fbxexportpreset
 fontconfig
+furtoolutils
+gasresizedynamic
 gplay
+groomingradial
 happrentice
 hconfig
 hdas
+hdefereval
 heducation
 hengine
+hjsonrpc
+hnac
+houcportion
 houdinicore
+houdiniengineutils
 houdinifx
+houdinihelp
+houdiniinternals
+houfs
+houxmlrpc
+hqrop
+hrpyc
+hscp
 htmlcov
 httpx
+husd
+husdui
+husktrace
+hutil
 hview
+hwebserver
 hython
 icccm
 infile
 ipynb
 ipython
+ishouenumtype
 jscoverage
 jspm
 keysyms
 keytar
 KHTML
+landmarking
 libasound
 libgl
 libglu
@@ -69,27 +129,91 @@ libxrender
 libxss
 libxt
 libxtst
+lightlinker
+lightmixer
+lightstate
 logstream
+logviewer
+lopcamandlightutils
+lopcamutils
+lopshaderutils
+loptoolutils
+loputils
+lpetags
 markserv
+materiallinker
 Microbundle
 mkdocs
 mmdc
+modelview
+msbuild
+mtlx
+muscletoolutils
+mvexportutils
 mypy
+ndarray
+nodegraph
+nodegraphalign
+nodegraphautoscroll
+nodegraphbase
+nodegraphconnect
+nodegraphdisplay
+nodegraphdispopts
+nodegraphedittext
+nodegraphfastfind
+nodegraphflags
+nodegraphfurutils
+nodegraphgestures
+nodegraphhotkeys
+nodegraphlayout
+nodegraphoptui
+nodegraphpalettes
+nodegraphpopmenus
+nodegraphprefs
+nodegraphquicknav
+nodegraphradialmenu
+nodegraphrename
+nodegraphsearch
+nodegraphselecthooks
+nodegraphselectpos
+nodegraphsnap
+nodegraphstates
+nodegraphtitle
+nodegraphutils
+nodegraphvellumutils
+nodegraphview
+nodeselectionutil
+nodethemes
 nosetests
 nuxt
 nvmrc
+objecttoolutils
+objtranslator
+ocio
 oudini
+parsepeakuserlog
+particletoolutils
+pdgd
+pdgdatalayer
+pdgjob
+pdgpathmap
+pdgservicepanel
+pdgutils
+perfmon
 pids
 pipenv
 Pipfile
 platbase
 platlib
+posespacedeform
+ppdgstateserver
 procps
 psutil
 purelib
 pybuilder
 pycache
 pyenv
+pyexpat
 pyflow
 pypa
 pypackages
@@ -100,32 +224,79 @@ pytest
 PYTHONDONTWRITEBYTECODE
 PYTHONPATH
 PYTHONPLATLIBDIR
+pythonscriptmenu
 PYTHONVERBOSE
 pytype
 pyvenv
 pyver
+radialmenu
 randr
+recipetoolutils
 Referer
+renderstats
+rendertracker
+resourceui
+rigtoolutils
+rmands
 ropeproject
+roptoolutils
+roputils
 scrapy
+scenegraphdetails
+scenegraphlayers
+sceneviewerhooks
 sdata
 sdist
 setuptools
+shaderhda
+shadingutils
+shelfutils
 shellquote
+shiboken
 shlvl
+shopclerks
+shoptoolutils
 SIDEFX
+simtracker
 sitecustomize
+skytoolutils
+snippetmenu
+soptoolutils
+soputils
 Spyder
 spyderproject
 spyproject
+stagemanager
+statehud
+stroketoolutils
 stylelintcache
 subprojeccts
 subproject
 SUBPROJECTS
+taskgraphtable
+taskgraphtablepanel
+terraintoolutils
+tilefx
 tomlkit
 toolbelt
+toolprompts
+toolutils
+topnettoolutils
+toptoolutils
+uiutils
+usdinlinelayermenu
+usdprimicons
+usdrenderers
+usdroputils
 uvicorn
+viewerhandle
+viewerstate
 vitepress
+volumetoolutils
+vopcallbacks
+vopfxmenu
+vopnettoolutils
+voptoolutils
 vscodeignore
 vscpp
 vuepress
