+++ houdini/h20.5/.vscode/launch.json
@@ -1,0 +1,20 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+
+
+
+        {
+            "name": "Python: Current File (h20-5)",
+            "type": "debugpy",
+            "request": "launch",
+            "program": "${file}",
+            "console": "integratedTerminal",
+            "python": "/Applications/Houdini/Houdini20.5.584/Frameworks/Houdini.framework/Resources/bin/hython",
+            "env": {
+                "PYTHONPATH": "${workspaceFolder:h20.5}/src${pathSeparator}${env:PYTHONPATH}"
+            },
+            "args": ["${command:pickArgs}"],
+        }
+    ]
+}
