+++ houdini/h20.5/.vscode/settings.json
@@ -1,2 +1,27 @@
 {
+    "files.exclude": {
+        "**/.ruby-lsp": true,
+        "zabob-chat/": true,
+        "mcp-server/": true,
+        "houdini/h*.*/": true,
+        "houdini/hdas/": true,
+        "dist": true,
+        "out": true,
+        "node_modules": true,
+        "zabob-modules/": true,
+        "docs/": true,
+        "docker/": true,
+        "*/docs/": true,
+        ".vscode-test/": true,
+        ".vscode-test-insiders/": true,
+        "**/__pycache__": true,
+        "**/*.egg-info": true,
+        "*~": true,
+        "*.bak": true,
+        "#*": true,
+        "*#": true
+    },
+    "python.analysis.extraPaths": [
+        "/Applications/Houdini/Houdini20.5.584/Frameworks/Houdini.framework/Resources/houdini/python3.11libs"
+    ]
 }
