+++ houdini/h20.5/pyproject.toml
@@ -9,6 +9,7 @@ authors = [
 requires-python = ">=3.11"
 dependencies = [
     "hou>=0.1.0",
+    "numpy>=2.2.6",
 ]
 use-hython = true
 
@@ -18,3 +19,6 @@ h20-5 = "h20_5:main"
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+include = [ "src/*" ]
