+++ mcp-server/pyproject.toml
@@ -9,7 +9,6 @@ authors = [
 requires-python = ">=3.13"
 dependencies = [
     "aiohttp>=3.11.18",
-    "aiopath>=0.7.7",
     "anyio>=4.9.0",
     "click>=8.1.8",
     "fastapi>=0.115.12",
@@ -21,6 +20,7 @@ dependencies = [
     "semver>=3.0.4",
     "sqlite-vec>=0.1.7a2",
     "uvicorn>=0.34.2",
+    "aiopath>=0.7.7",
 ]
 
 [build-system]
