+++ mcp-server/src/zsbob/mcp/instructions.md
