+++ mcp-server/src/zsbob/mcp/responses/list_node_types.md
