+++ mcp-server/src/zsbob/mcp/server.py
