+++ zabob-modules/src/zabob/core/houdini.py
@@ -16,12 +16,12 @@ from semver import Version
 
 
 from zabob._find.types import HoudiniInstall
-from zabob.houdini_versions import cli as houdini_cli, download_houdini_installer
-from zabob.find_houdini import get_houdini, list_houdini_installations, show_houdini
-from zabob.main import main
-from zabob.paths import HOUDINI_PROJECTS, ZABOB_HOUDINI_DIR
-from zabob.subproc import run
-from zabob.hython import hython
+from zabob.core.houdini_versions import cli as houdini_cli, download_houdini_installer
+from zabob.core.find_houdini import get_houdini, list_houdini_installations, show_houdini
+from zabob.core.main import main
+from zabob.core.paths import HOUDINI_PROJECTS, ZABOB_HOUDINI_DIR
+from zabob.core.subproc import run
+from zabob.core.hython import hython
 
 @main.group('houdini')
 def houdini_commands():
