+++ zabob-modules/src/zabob/core/paths.py
